+++ jflu-producer-rabbitmq-quarkus/src/main/java/be/looorent/jflu/publisher/rabbitmq/quarkus/ProducerRecorder$accessor.java
@@ -1,8 +1,0 @@
-package be.looorent.jflu.publisher.rabbitmq.quarkus;
-
-public final class ProducerRecorder$accessor {
-    private ProducerRecorder$accessor() {}
-    public static Object construct() {
-        return new ProducerRecorder();
-    }
-}