+++ jflu-producer-rabbitmq/src/main/java/be/looorent/jflu/publisher/rabbitmq/RabbitMQConnectionFactory.java
@@ -10,7 +10,6 @@ import java.security.KeyManagementException;
 import java.security.NoSuchAlgorithmException;
 import java.util.Properties;
 import java.util.concurrent.TimeoutException;
-import java.util.function.Predicate;
 
 import static be.looorent.jflu.publisher.rabbitmq.RabbitMQPropertyName.*;
 import static java.lang.Thread.sleep;
@@ -20,7 +19,7 @@ import static java.util.Optional.ofNullable;
  * Attempt to connect RabbitMQ and wait if needed.
  * @author Lorent Lempereur {@literal <lorent.lempereur.dev@gmail.com>}
  */
-class RabbitMQConnectionFactory {
+public class RabbitMQConnectionFactory {
 
     private static final Logger LOG = LoggerFactory.getLogger(RabbitMQConnectionFactory.class);
     private static final int MAXIMUM_CONNECTION_ATTEMPTS = 30;