+++ jflu-subscriber-rabbitmq-quarkus-deployment/src/main/java/be/looorent/jflu/publisher/rabbitmq/quarkus/deployment/JFluSubscriberProcessor$accessor.java
@@ -1,8 +1,0 @@
-package be.looorent.jflu.publisher.rabbitmq.quarkus.deployment;
-
-public class JFluSubscriberProcessor$accessor {
-    private JFluSubscriberProcessor$accessor() {}
-    public static Object construct() {
-        return new JFluSubscriberProcessor();
-    }
-}