+++ jflu-subscriber-rabbitmq-quarkus/src/main/java/be/looorent/jflu/publisher/rabbitmq/quarkus/SubscriberRecorder$accessor.java
@@ -1,8 +1,0 @@
-package be.looorent.jflu.publisher.rabbitmq.quarkus;
-
-public class SubscriberRecorder$accessor {
-    private SubscriberRecorder$accessor() {}
-    public static Object construct() {
-        return new SubscriberRecorder();
-    }
-}