+++ jflu-subscriber-rabbitmq-quarkus/src/main/java/be/looorent/jflu/publisher/rabbitmq/quarkus/SubscriberRuntimeConfiguration$accessor.java
@@ -1,8 +1,0 @@
-package be.looorent.jflu.publisher.rabbitmq.quarkus;
-
-public class SubscriberRuntimeConfiguration$accessor {
-    private SubscriberRuntimeConfiguration$accessor() {}
-    public static Object construct() {
-        return new SubscriberRuntimeConfiguration();
-    }
-}