+++ CHANGELOG.md
@@ -5,10 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.10] - 2022-08-10
+## [1.0.12] - 2022-08-10
 
 * Include Jandex files in artifacts
 * Increase Quarkus Beans discoverability
+* Use Gradle to build the Quarkus extensions
+* Extension for RabbitMQ Subscription (`RabbitMQSubscriptionBootstraper`)
 
 ## [1.0.8] - 2022-08-09